+++ modules/api.js
@@ -18,4 +18,13 @@ async function getData(name) {
     return meds
 }
 
+function textCompressed(medsName) {
+    let name = medsName
+    console.log(name)
+    let str = name.substr(0, 30)
+    return str
+}
+
+// textCompressed('LevonorgestrelV/Ethinylestradiol Aurobi o 3X27 filmomhulde tabletten ‘ Bevat 0.a. lactosemonohydraat. Zie bijsluiter voor verdere informatie. L ees voor het gebruik de bijsluiter. Voor oraai gebruik. guiten het zicht en bereik van kinderen houden. gewaren beneden 30°C. RYG 111871 UR.')
+
 module.exports = getData
